backend/pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ncert-class6-study-backend"
version = "0.1.0"
description = "FastAPI backend for an NCERT Class 6 intelligent study assistant"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "pydantic>=2.8.0",
  "pydantic-settings>=2.4.0",
  "langchain>=0.3.27,<0.4",
  "langchain-community>=0.3.27,<0.4",
  "langchain-core>=0.3.66,<1.0",
  "langchain-text-splitters>=0.3.11,<0.4",
  "langchain-ollama>=0.2.3,<0.3",
  "langchain-google-genai>=2.0.9,<3.0",
  "google-genai>=1.28.0,<2.0",
  "tiktoken>=0.7.0",
  "pymupdf>=1.24.0",
  "langchain-huggingface>=0.1.2,<0.2",
  "sentence-transformers>=3.0.0",
  "chromadb>=0.5.0",
  "pypdf>=4.3.0",
  "python-multipart>=0.0.9",
  "httpx>=0.27.0",
  "tenacity>=8.5.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.3.0", "pytest-asyncio>=0.24.0", "ruff>=0.6.0"]

[project.scripts]
ncert-backend = "app.main:run"

[tool.setuptools]
package-dir = {"" = "."}

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
